+++ library/Exceptions/AbstractGroupedException.php
@@ -15,6 +15,7 @@ class AbstractGroupedException extends AbstractNestedException
 {
     const NONE = 0;
     const SOME = 1;
+
     public static $defaultTemplates = [
         self::MODE_DEFAULT => [
             self::NONE => 'All of the required rules must pass for {{name}}',
@@ -29,32 +30,8 @@ class AbstractGroupedException extends AbstractNestedException
     public function chooseTemplate()
     {
         $numRules = $this->getParam('passed');
-        $numFailed = count($this->getRelated());
+        $numFailed = $this->getRelated()->count();
 
         return $numRules === $numFailed ? static::NONE : static::SOME;
-    }
-
-    public function getParams()
-    {
-        if (1 === count($this->related)) {
-            return current($this->related)->getParams();
-        } else {
-            return parent::getParams();
-        }
-    }
-
-    public function getTemplate()
-    {
-        $parentTemplate = parent::getTemplate();
-        $isEmpty = empty($this->template);
-
-        if (!$isEmpty && $this->template != $parentTemplate) {
-            return $this->template;
-        }
-        if ($isEmpty && 1 === count($this->related)) {
-            return current($this->related)->getTemplate();
-        } else {
-            return $parentTemplate;
-        }
     }
 }