+++ library/Exceptions/NestedValidationExceptionInterface.php
@@ -11,7 +11,9 @@
 
 namespace Respect\Validation\Exceptions;
 
-interface NestedValidationExceptionInterface extends ValidationExceptionInterface
+use IteratorAggregate;
+
+interface NestedValidationExceptionInterface extends IteratorAggregate, ValidationExceptionInterface
 {
     public function findMessages(array $paths);
     public function getMessages();